+++ src/main/java/pe/edu/cibertec/waS3ventacibertec/model/dto/request/ProductRequest.java
@@ -1,0 +1,14 @@
+package pe.edu.cibertec.waS3ventacibertec.model.dto.request;
+
+import lombok.Data;
+
+@Data
+public class ProductRequest {
+
+    private Integer productId;
+    private String productname;
+    private Double unitprice;
+    private Boolean discontinued;
+    private Integer supplierid;
+    private Integer categoryid;
+}
