+++ src/main/java/pe/edu/cibertec/waS3ventacibertec/repository/ProductRepository.java
@@ -1,4 +1,4 @@
-package pe.edu.cibertec.waS3ventacibertec.model.repository;
+package pe.edu.cibertec.waS3ventacibertec.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
