+++ src/main/java/pe/edu/cibertec/waS3ventacibertec/service/IProductService.java
@@ -1,4 +1,4 @@
-package pe.edu.cibertec.waS3ventacibertec.model.service;
+package pe.edu.cibertec.waS3ventacibertec.service;
 
 import pe.edu.cibertec.waS3ventacibertec.model.bd.Product;
